chore: move source code from applications/concept-catalogue to root and change package prefix to no.brreg.conceptcatalogue
