+++ buildCommands.sh
@@ -16,7 +16,7 @@ maven_apps="concept-catalogue"
 for maven_app in $maven_apps; do
     i=$((i+1))
     BUILD_APPS[$i]=$maven_app
-    BUILD_CMD[$i]="mvn clean install -B -T 2C --projects applications/$maven_app --also-make"
+    BUILD_CMD[$i]="mvn clean install -B -T 2C --also-make"
 done
 
 export BUILD_APPS
