+++ src/main/kotlin/no/brreg/conceptcatalogue/ConceptRegistrationApplication.kt
@@ -1,4 +1,4 @@
-package no.begrepskatalog.conceptregistration
+package no.brreg.conceptcatalogue
 
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
