+++ src/main/kotlin/no/brreg/conceptcatalogue/configuration/SecurityConfigurer.java
@@ -1,4 +1,4 @@
-package no.begrepskatalog.conceptregistration.configuration;
+package no.brreg.conceptcatalogue.configuration;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.security.oauth2.resource.ResourceServerProperties;
