+++ src/main/resources/static/ConceptRegistrationAPI.yaml
