+++ src/main/resources/static/skos-ap-no.yaml
