+++ src/main/resources/templates/openapi-generator-maven-plugin/templates/JavaSpring/bodyParams.mustache
